requirements.txt
djangorestframework>=3.10.0,<3.11.0
django>=1.11.0,<1.12.0
ipython>=7.1.0,<7.2.0

pillow>=6.1.0,<6.2.0
flake8>=3.7.0,<3.8.0
